pyproject.toml
[tool.poetry]
name = "ASEDailyStartUp"
version = "0.1.0"
description = "Internal Application for Learnosity Applicatiojn Support Engineering Team"
authors = ["Terry A. Brooks, Jr."]
license = "MIY"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
redis = "^5.0.1"
hiredis = "^2.2.3"
gunicorn = "^21.2.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"
logtail-python = "^0.2.6"
loguru = "^0.7.2"
psycopg2-binary = "^2.9.9"
django-health-check = "^3.17.0"
python-dotenv = "^1.0.0"
django-filter = "^23.3"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
whitenoise = "^6.5.0"
martor = "^1.6.28"
django-cors-headers = "^4.3.0"
arrow = "^1.3.0"
django-debug-toolbar = "^4.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pre-commit = "^3.4.0"
build = "^1.0.3"
djhtml = "^3.0.6"
kolo = "^2.16.3"
autoflake = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.djlint]
format_css=true
format_js=true
ignore="H021,T001, H008, H017"

[tool.djlint.css]
indent_size=5

[tool.djlint.js]
indent_size=5

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
